removed the pokemon stuff. was distracting
